refactor: extract publishing shipping form section
