+++ src/kleinanzeigen_bot/resources/translations.de.yaml
@@ -141,9 +141,6 @@ kleinanzeigen_bot/__init__.py:
     "Post-publish persistence failed for '%s' (ad ID %s - ad is live on Kleinanzeigen but local YAML may be out of sync)": "Persistenz nach Veröffentlichung fehlgeschlagen für '%s' (Anzeigen-ID %s - Anzeige ist auf Kleinanzeigen live, aber die lokale YAML ist möglicherweise nicht synchron)"
 
   _fill_ad_form:
-    "Failed to set shipping attribute for type '%s'!": "Fehler beim setzen des Versandattributs für den Typ '%s'!"
-    "Shipping step skipped - reason: NOT_APPLICABLE": "Versandschritt übersprungen: Versand nicht anwendbar (Status = NOT_APPLICABLE)"
-    "Shipping combobox button has no id attribute": "Versand-Combobox-Button hat kein ID-Attribut"
 
   update_ads:
     "Processing %s/%s: '%s' from [%s]...": "Verarbeite %s/%s: '%s' von [%s]..."
@@ -165,9 +162,6 @@ kleinanzeigen_bot/__init__.py:
     "Unable to select condition [%s]": "Zustand [%s] konnte nicht ausgewählt werden"
     "Failed to set attribute '%s'": "Fehler beim Setzen des Attributs '%s'"
     "No condition radio matched values %(values)s": "Keine Zustand-Auswahl passt zu den Werten %(values)s"
-
-  _select_button_combobox:
-    "Option '%(value)s' not found in button combobox '%(id)s'": "Option '%(value)s' nicht in Button-Combobox '%(id)s' gefunden"
 
   _set_special_attributes:
     "Found %i special attributes": "%i spezielle Attribute gefunden"
@@ -196,25 +190,6 @@ kleinanzeigen_bot/__init__.py:
     "DONE: No ads found to extend.": "FERTIG: Keine Anzeigen zum Verlängern gefunden."
     "Unknown command: %s": "Unbekannter Befehl: %s"
     "Timing collector flush failed: %s": "Zeitmessdaten konnten nicht gespeichert werden: %s"
-
-  _set_shipping:
-    "Failed to set shipping attribute for type '%s'!": "Fehler beim setzen des Versandattributs für den Typ '%s'!"
-    "Shipping combobox button has no id attribute": "Versand-Combobox-Button hat kein ID-Attribut"
-    "Shipping fieldset is rendered, but the pickup radio is missing; page may not be fully loaded.": "Das Versand-Fieldset wird gerendert, aber die Abholungs-Option fehlt; die Seite wurde möglicherweise nicht vollständig geladen."
-    "Unsupported shipping_type: %s": "Nicht unterstützter shipping_type: %s"
-    ? "Shipping options dialog entry not found. Legacy '.versand_s' select UI is no longer supported and requires dedicated rebuild."
-    : "Einstieg in den Versandoptionen-Dialog nicht gefunden. Die alte '.versand_s'-Select-UI wird nicht mehr unterstützt und muss gezielt neu umgesetzt werden."
-    "Unable to open shipping options dialog!": "Versandoptionen-Dialog konnte nicht geöffnet werden!"
-    "Unable to select individual shipping option!": "Individuelle Versandoption konnte nicht ausgewählt werden!"
-    "Unable to set shipping price!": "Versandpreis konnte nicht gesetzt werden!"
-    "Unable to close shipping dialog!": "Versanddialog konnte nicht geschlossen werden!"
-
-  _set_shipping_options:
-    "Failed to configure shipping options in dialog!": "Versandoptionen konnten im Dialog nicht konfiguriert werden!"
-    "Unable to close shipping dialog!": "Versanddialog konnte nicht geschlossen werden!"
-    "shipping_options must be provided": "shipping_options müssen angegeben werden!"
-    "Unknown shipping option(s), please refer to the documentation/README: %s": "Unbekannte Versandoption(en), bitte lies die Dokumentation/das README: %s"
-    "You can only specify shipping options for one package size!": "Du kannst nur Versandoptionen für eine Paketgröße angeben!"
 
 #################################################
 kleinanzeigen_bot/published_ads.py:
@@ -281,6 +256,33 @@ kleinanzeigen_bot/publishing_form.py:
 
   check_thumbnails_uploaded:
     " -> %d of %d images processed": " -> %d von %d Bildern verarbeitet"
+
+  set_shipping_form:
+    "Failed to set shipping attribute for type '%s'!": "Fehler beim setzen des Versandattributs für den Typ '%s'!"
+    "Shipping step skipped - reason: NOT_APPLICABLE": "Versandschritt übersprungen: Versand nicht anwendbar (Status = NOT_APPLICABLE)"
+    "Shipping combobox button has no id attribute": "Versand-Combobox-Button hat kein ID-Attribut"
+
+  _select_button_combobox:
+    "Option '%(value)s' not found in button combobox '%(id)s'": "Option '%(value)s' nicht in Button-Combobox '%(id)s' gefunden"
+
+  set_shipping:
+    "Failed to set shipping attribute for type '%s'!": "Fehler beim setzen des Versandattributs für den Typ '%s'!"
+    "Shipping combobox button has no id attribute": "Versand-Combobox-Button hat kein ID-Attribut"
+    "Shipping fieldset is rendered, but the pickup radio is missing; page may not be fully loaded.": "Das Versand-Fieldset wird gerendert, aber die Abholungs-Option fehlt; die Seite wurde möglicherweise nicht vollständig geladen."
+    "Unsupported shipping_type: %s": "Nicht unterstützter shipping_type: %s"
+    ? "Shipping options dialog entry not found. Legacy '.versand_s' select UI is no longer supported and requires dedicated rebuild."
+    : "Einstieg in den Versandoptionen-Dialog nicht gefunden. Die alte '.versand_s'-Select-UI wird nicht mehr unterstützt und muss gezielt neu umgesetzt werden."
+    "Unable to open shipping options dialog!": "Versandoptionen-Dialog konnte nicht geöffnet werden!"
+    "Unable to select individual shipping option!": "Individuelle Versandoption konnte nicht ausgewählt werden!"
+    "Unable to set shipping price!": "Versandpreis konnte nicht gesetzt werden!"
+    "Unable to close shipping dialog!": "Versanddialog konnte nicht geschlossen werden!"
+
+  set_shipping_options:
+    "Failed to configure shipping options in dialog!": "Versandoptionen konnten im Dialog nicht konfiguriert werden!"
+    "Unable to close shipping dialog!": "Versanddialog konnte nicht geschlossen werden!"
+    "shipping_options must be provided": "shipping_options müssen angegeben werden!"
+    "Unknown shipping option(s), please refer to the documentation/README: %s": "Unbekannte Versandoption(en), bitte lies die Dokumentation/das README: %s"
+    "You can only specify shipping options for one package size!": "Du kannst nur Versandoptionen für eine Paketgröße angeben!"
 
 #################################################
 kleinanzeigen_bot/publishing_persistence.py:
