+++ tests/unit/test_publishing_form.py
@@ -13,7 +13,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 
 from kleinanzeigen_bot import KleinanzeigenBot
-from kleinanzeigen_bot.model.ad_model import Ad
+from kleinanzeigen_bot.ad_form_helpers import VERSAND_COMBOBOX_SELECTOR
+from kleinanzeigen_bot.model.ad_model import Ad, AdUpdateStrategy
 from kleinanzeigen_bot.publishing_form import (
     city_option_text,
     fill_image_section,
@@ -24,6 +25,9 @@ from kleinanzeigen_bot.publishing_form import (
     set_contact_fields,
     set_contact_location,
     set_pricing_fields,
+    set_shipping,
+    set_shipping_form,
+    set_shipping_options,
     upload_images,
 )
 from kleinanzeigen_bot.utils.exceptions import CategoryResolutionError
@@ -1258,3 +1262,616 @@ class TestPricingFields:
 
         mock_desc.assert_called_once_with(ad_cfg, test_bot.config.ad_defaults, with_affixes = True)
         mock_set.assert_any_await("ad-description", "Expected description text")
+
+
+class TestShippingDialogFlow:
+    """Regression tests for shipping dialog flow using new radio selectors only."""
+
+    shipping_combobox_selector = VERSAND_COMBOBOX_SELECTOR
+
+    @pytest.mark.asyncio
+    @pytest.mark.parametrize(
+        ("shipping_type", "expected_label"),
+        [("SHIPPING", "Versand möglich"), ("PICKUP", "Nur Abholung")],
+    )
+    async def test_shipping_uses_versand_combobox_when_rendered(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+        shipping_type:str,
+        expected_label:str,
+    ) -> None:
+        """Commercial accounts may render Versand as a special-attribute combobox instead of radio buttons."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": shipping_type})
+        shipping_combobox = MagicMock()
+        shipping_combobox.attrs = {"id": "uhren.versand"}
+
+        with (
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = shipping_combobox) as mock_probe,
+            patch.object(test_bot, "web_select_button_combobox", new_callable = AsyncMock) as mock_select_combobox,
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        mock_probe.assert_awaited_once()
+        assert mock_probe.await_args is not None
+        assert mock_probe.await_args.args[:2] == (
+            By.CSS_SELECTOR,
+            self.shipping_combobox_selector,
+        )
+        mock_select_combobox.assert_awaited_once()
+        assert mock_select_combobox.await_args is not None
+        assert mock_select_combobox.await_args.args[:2] == ("uhren.versand", expected_label)
+        mock_click.assert_not_awaited()
+
+    @pytest.mark.asyncio
+    @pytest.mark.parametrize("selected", [False, True])
+    async def test_pickup_shipping_radio_selection(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+        selected:bool,
+    ) -> None:
+        """PICKUP shipping should click the pickup radio only when it is not already selected."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "PICKUP"})
+
+        with (
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, side_effect = [None, MagicMock()]) as mock_probe,
+            patch.object(test_bot, "web_check", new_callable = AsyncMock, return_value = selected),
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        observed = [call.args[:2] for call in mock_probe.await_args_list]
+        assert (By.CSS_SELECTOR, self.shipping_combobox_selector) in observed
+        assert (By.ID, "ad-shipping-enabled-no") in observed
+        if selected:
+            mock_click.assert_not_awaited()
+        else:
+            mock_click.assert_awaited_once()
+            assert mock_click.call_args.args[:2] == (By.ID, "ad-shipping-enabled-no")
+
+    @pytest.mark.asyncio
+    async def test_pickup_shipping_raises_when_radio_lookup_times_out(self, test_bot:KleinanzeigenBot, base_ad_config:dict[str, Any]) -> None:
+        """PICKUP shipping should fail fast when pickup radio selector is unavailable."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "PICKUP"})
+
+        with (
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, side_effect = [None, MagicMock()]),
+            patch.object(test_bot, "web_check", new_callable = AsyncMock, side_effect = TimeoutError("pickup lookup timed out")),
+            pytest.raises(TimeoutError, match = "Failed to set shipping attribute for type 'PICKUP'!"),
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+    @pytest.mark.asyncio
+    async def test_pickup_shipping_skips_when_toggle_not_rendered(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """Categories without a shipping fieldset (e.g. books 76/77, comics 76/77/15156)
+        are PICKUP-only by site convention — the absence of both shipping selectors should
+        short-circuit without calling ``web_check``/``web_click``."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "PICKUP"})
+
+        with (
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, side_effect = [None, None, None]),
+            patch.object(test_bot, "web_check", new_callable = AsyncMock) as mock_check,
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        mock_check.assert_not_awaited()
+        mock_click.assert_not_awaited()
+
+    @pytest.mark.asyncio
+    async def test_pickup_shipping_raises_when_fieldset_rendered_but_pickup_radio_missing(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """A rendered shipping fieldset without the pickup radio should be treated as an error."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "PICKUP"})
+
+        with (
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, side_effect = [None, None, MagicMock()]) as mock_probe,
+            patch.object(test_bot, "web_check", new_callable = AsyncMock) as mock_check,
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            pytest.raises(
+                TimeoutError,
+                match = "Shipping fieldset is rendered, but the pickup radio is missing; page may not be fully loaded.",
+            ),
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        observed = [call.args[:2] for call in mock_probe.await_args_list]
+        assert (By.CSS_SELECTOR, self.shipping_combobox_selector) in observed
+        assert (By.ID, "ad-shipping-enabled-no") in observed
+        assert (By.ID, "ad-shipping-enabled") in observed
+        mock_check.assert_not_awaited()
+        mock_click.assert_not_awaited()
+
+    @pytest.mark.asyncio
+    async def test_shipping_without_options_uses_radio_and_dialog(self, test_bot:KleinanzeigenBot, base_ad_config:dict[str, Any]) -> None:
+        """Shipping without package options should use radio + dialog flow."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "SHIPPING", "shipping_options": [], "shipping_costs": 4.95})
+
+        with (
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None),
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = MagicMock()),
+            patch.object(test_bot, "web_set_input_value", new_callable = AsyncMock) as mock_set_input,
+            patch.object(test_bot, "web_execute", new_callable = AsyncMock, return_value = "4,95"),
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+            click_args = [c.args for c in mock_click.await_args_list]
+            assert any(len(a) >= 2 and a[0] == By.ID and a[1] == "ad-individual-shipping-checkbox-control" for a in click_args)
+            assert any("Fertig" in str(a[1]) for a in click_args if len(a) >= 2)
+            mock_set_input.assert_awaited_once_with("ad-individual-shipping-price", "4,95")
+
+    @pytest.mark.asyncio
+    async def test_shipping_finish_timeout_raises(self, test_bot:KleinanzeigenBot, base_ad_config:dict[str, Any]) -> None:
+        """Timeout while confirming shipping dialog should raise a clear error."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "SHIPPING", "shipping_options": []})
+
+        async def click_side_effect(selector_type:By, selector_value:str, **_:Any) -> None:
+            if selector_type == By.XPATH and "Fertig" in selector_value:
+                raise TimeoutError("finish timeout")
+
+        with (
+            patch.object(test_bot, "web_click", new_callable = AsyncMock, side_effect = click_side_effect),
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None),
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = MagicMock()),
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+            pytest.raises(TimeoutError, match = "Unable to close shipping dialog!"),
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+    @pytest.mark.asyncio
+    async def test_shipping_without_options_does_not_toggle_checkbox_when_price_input_visible(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """When price input is already visible, individual-shipping checkbox is not toggled."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "SHIPPING", "shipping_options": [], "shipping_costs": 4.95})
+
+        with (
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, side_effect = [None, MagicMock()]),
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = MagicMock()),
+            patch.object(test_bot, "web_set_input_value", new_callable = AsyncMock) as mock_set_input,
+            patch.object(test_bot, "web_execute", new_callable = AsyncMock, return_value = "4,95"),
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        click_args = [c.args for c in mock_click.await_args_list]
+        assert not any(len(a) >= 2 and a[0] == By.ID and a[1] == "ad-individual-shipping-checkbox-control" for a in click_args)
+        mock_set_input.assert_awaited_once_with("ad-individual-shipping-price", "4,95")
+
+    @pytest.mark.asyncio
+    async def test_shipping_price_lost_to_react_rerender_raises(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """When React re-render swallows the shipping price, the dialog must NOT be closed."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "SHIPPING", "shipping_options": [], "shipping_costs": 4.95})
+
+        async def set_input_side_effect(element_id:str, value:str) -> None:
+            pass
+
+        with (
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None),
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = MagicMock()),
+            patch.object(test_bot, "web_set_input_value", new_callable = AsyncMock, side_effect = set_input_side_effect),
+            patch.object(test_bot, "web_execute", new_callable = AsyncMock, return_value = None),
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+            pytest.raises(TimeoutError, match = "Unable to set shipping price!"),
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        # Fertig must never be clicked when the price was not confirmed
+        fertig_clicks = [c for c in mock_click.await_args_list if c.args and len(c.args) >= 2 and "Fertig" in str(c.args[1])]
+        assert not fertig_clicks, "Fertig was clicked despite shipping price not being set"
+
+    @pytest.mark.asyncio
+    async def test_shipping_price_recovers_when_readback_matches_on_retry(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """First attempt mismatches (readback empty), second attempt succeeds — must close the dialog normally."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "SHIPPING", "shipping_options": [], "shipping_costs": 4.95})
+
+        with (
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None),
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = MagicMock()),
+            patch.object(test_bot, "web_set_input_value", new_callable = AsyncMock) as mock_set_input,
+            patch.object(test_bot, "web_execute", new_callable = AsyncMock, side_effect = [None, "4,95"]),
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_sleep,
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        assert mock_set_input.await_count == 2, "expected exactly one retry after the first mismatch"
+        inter_attempt_sleeps = [c for c in mock_sleep.await_args_list if c.args == (300, 500)]
+        assert len(inter_attempt_sleeps) == 1, "expected one inter-attempt backoff sleep"
+        fertig_clicks = [c for c in mock_click.await_args_list if c.args and len(c.args) >= 2 and "Fertig" in str(c.args[1])]
+        assert fertig_clicks, "Fertig must be clicked once the readback confirms the price"
+
+    @pytest.mark.asyncio
+    async def test_shipping_price_retries_when_readback_raises_transiently(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """A TimeoutError from the readback web_execute must be retried, not propagated on the first occurrence."""
+        ad_cfg = Ad.model_validate(base_ad_config | {"shipping_type": "SHIPPING", "shipping_options": [], "shipping_costs": 4.95})
+
+        readback_results:list[str | Exception] = [TimeoutError("readback raced with re-render"), "4,95"]
+
+        async def readback_side_effect(*_args:Any, **_kwargs:Any) -> str | None:
+            result = readback_results.pop(0)
+            if isinstance(result, Exception):
+                raise result
+            return result
+
+        with (
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None),
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = MagicMock()),
+            patch.object(test_bot, "web_set_input_value", new_callable = AsyncMock) as mock_set_input,
+            patch.object(test_bot, "web_execute", new_callable = AsyncMock, side_effect = readback_side_effect),
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+        ):
+            await set_shipping(test_bot, ad_cfg)
+
+        assert mock_set_input.await_count == 2, "expected one retry after the readback raised"
+        fertig_clicks = [c for c in mock_click.await_args_list if c.args and len(c.args) >= 2 and "Fertig" in str(c.args[1])]
+        assert fertig_clicks, "Fertig must be clicked once a later readback confirms the price"
+
+
+class TestShippingOptionsDialog:
+    """Tests for set_shipping_options using carrier-code-based selectors."""
+
+    @staticmethod
+    def _make_ad_with_options(base_ad_config:dict[str, Any], options:list[str]) -> Ad:
+        return Ad.model_validate(
+            base_ad_config
+            | {
+                "shipping_type": "SHIPPING",
+                "shipping_options": options,
+            }
+        )
+
+    @staticmethod
+    def _mock_checkbox(checked:bool = False) -> MagicMock:
+        """Create a mock checkbox element with optional checked attribute."""
+        el = MagicMock()
+        if checked:
+            el.attrs = {"checked": ""}
+        else:
+            el.attrs = {}
+        return el
+
+    @pytest.mark.parametrize(
+        "case",
+        [
+            # SMALL pre-checked, only unwanted carriers are toggled
+            {
+                "options": ["Hermes_Päckchen"],
+                "radio_checked": True,
+                "expected_radio_click": False,
+                "expected_clicked_carriers": ["HERMES_002", "DHL_001"],
+                "expected_not_clicked_carriers": ["HERMES_001"],
+            },
+            # LARGE not checked, radio click needed and only unwanted carriers are toggled
+            {
+                "options": ["DHL_10"],
+                "radio_checked": False,
+                "expected_radio_click": True,
+                "expected_clicked_carriers": ["HERMES_004", "DHL_004", "DHL_005"],
+                "expected_not_clicked_carriers": ["DHL_003"],
+            },
+        ],
+    )
+    @pytest.mark.asyncio
+    async def test_replace_mode_handles_radio_state(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+        case:dict[str, Any],
+    ) -> None:
+        """REPLACE mode: handles both pre-checked and unchecked radio states."""
+        ad_cfg = self._make_ad_with_options(base_ad_config, case["options"])
+
+        radio_mock = self._mock_checkbox(checked = case["radio_checked"])
+
+        async def find_side_effect(selector_type:By, selector_value:str, **_:Any) -> MagicMock:
+            if "radio" in selector_value:
+                return radio_mock
+            return self._mock_checkbox(checked = True)  # all checkboxes pre-checked
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, side_effect = find_side_effect),
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+        ):
+            await set_shipping_options(test_bot, ad_cfg, mode = AdUpdateStrategy.REPLACE)
+
+        click_args = [(c.args[0], c.args[1]) for c in mock_click.await_args_list if len(c.args) >= 2]
+
+        # Radio click behavior matches expectation
+        radio_clicked = any("radio" in str(a[1]) for a in click_args)
+        assert radio_clicked == case["expected_radio_click"]
+
+        # Should click Weiter and Fertig
+        assert any("Weiter" in str(a[1]) for a in click_args)
+        assert any("Fertig" in str(a[1]) for a in click_args)
+
+        # Should toggle exactly the expected carriers for this scenario
+        for carrier_code in case["expected_clicked_carriers"]:
+            assert any(carrier_code in str(a[1]) for a in click_args)
+
+        for carrier_code in case["expected_not_clicked_carriers"]:
+            assert not any(carrier_code in str(a[1]) for a in click_args)
+
+    @pytest.mark.asyncio
+    async def test_replace_mode_dom_verified_unchecked_defaults_select_wanted_carrier(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """REPLACE mode must select wanted carriers when defaults are unchecked (DOM-verified for MEDIUM/LARGE)."""
+        ad_cfg = self._make_ad_with_options(base_ad_config, ["DHL_5"])
+
+        radio_mock = self._mock_checkbox(checked = False)  # MEDIUM radio not selected yet
+
+        async def find_side_effect(selector_type:By, selector_value:str, **_:Any) -> MagicMock:
+            if "radio" in selector_value and "MEDIUM" in selector_value:
+                return radio_mock
+            # DOM probe confirms MEDIUM defaults can be unchecked after "Weiter"
+            if "HERMES_003" in selector_value:
+                return self._mock_checkbox(checked = False)
+            if "DHL_002" in selector_value:
+                return self._mock_checkbox(checked = False)
+            return self._mock_checkbox(checked = False)
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, side_effect = find_side_effect),
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+        ):
+            await set_shipping_options(test_bot, ad_cfg, mode = AdUpdateStrategy.REPLACE)
+
+        click_args = [(c.args[0], c.args[1]) for c in mock_click.await_args_list if len(c.args) >= 2]
+
+        # Regression guard for issue #956: wanted DHL_002 must be selected
+        assert any("DHL_002" in str(a[1]) for a in click_args)
+        # Unwanted Hermes checkbox must remain untouched when already unchecked
+        assert not any("HERMES_003" in str(a[1]) for a in click_args)
+
+    @pytest.mark.asyncio
+    async def test_modify_mode_toggles_carriers(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """MODIFY mode: explicitly (de-)selects each carrier based on wanted set."""
+        ad_cfg = self._make_ad_with_options(base_ad_config, ["Hermes_Päckchen", "DHL_2"])
+
+        radio_mock = self._mock_checkbox(checked = True)  # SMALL already selected
+
+        async def find_side_effect(selector_type:By, selector_value:str, **_:Any) -> MagicMock:
+            if "radio" in selector_value and "SMALL" in selector_value:
+                return radio_mock
+            # HERMES_001 checked, HERMES_002 checked, DHL_001 unchecked
+            if "HERMES_001" in selector_value:
+                return self._mock_checkbox(checked = True)
+            if "HERMES_002" in selector_value:
+                return self._mock_checkbox(checked = True)
+            if "DHL_001" in selector_value:
+                return self._mock_checkbox(checked = False)
+            return self._mock_checkbox(checked = False)
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, side_effect = find_side_effect),
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+        ):
+            await set_shipping_options(test_bot, ad_cfg, mode = AdUpdateStrategy.MODIFY)
+
+        click_args = [(c.args[0], c.args[1]) for c in mock_click.await_args_list if len(c.args) >= 2]
+        # HERMES_002 should be deselected (was checked, not wanted)
+        assert any("HERMES_002" in str(a[1]) for a in click_args)
+        # DHL_001 should be selected (was unchecked, wanted via DHL_2 → DHL_001)
+        assert any("DHL_001" in str(a[1]) for a in click_args)
+        # HERMES_001 should NOT be clicked (was checked, wanted)
+        assert not any("HERMES_001" in str(a[1]) for a in click_args)
+
+    @pytest.mark.asyncio
+    async def test_unknown_option_raises_key_error(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """Unknown shipping option name raises KeyError with helpful message."""
+        ad_cfg = self._make_ad_with_options(base_ad_config, ["NonExistent_Option"])
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_sleep,
+            pytest.raises(KeyError, match = "Unknown shipping option"),
+        ):
+            await set_shipping_options(test_bot, ad_cfg)
+
+        # Validation errors must occur before any DOM interaction
+        mock_find.assert_not_awaited()
+        mock_click.assert_not_awaited()
+        mock_sleep.assert_not_awaited()
+
+    @pytest.mark.asyncio
+    async def test_mixed_size_options_raises_value_error(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """Options from different size groups raise ValueError."""
+        ad_cfg = self._make_ad_with_options(base_ad_config, ["Hermes_Päckchen", "DHL_5"])
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
+            patch.object(test_bot, "web_click", new_callable = AsyncMock) as mock_click,
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_sleep,
+            pytest.raises(ValueError, match = "one package size"),
+        ):
+            await set_shipping_options(test_bot, ad_cfg)
+
+        # Validation errors must occur before any DOM interaction
+        mock_find.assert_not_awaited()
+        mock_click.assert_not_awaited()
+        mock_sleep.assert_not_awaited()
+
+    @pytest.mark.asyncio
+    async def test_timeout_in_dialog_raises(self, test_bot:KleinanzeigenBot, base_ad_config:dict[str, Any]) -> None:
+        """TimeoutError during dialog interaction is re-raised with descriptive message."""
+        ad_cfg = self._make_ad_with_options(base_ad_config, ["Hermes_Päckchen"])
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, side_effect = TimeoutError("radio not found")),
+            patch.object(test_bot, "web_click", new_callable = AsyncMock),
+            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
+            pytest.raises(TimeoutError, match = "Failed to configure shipping options in dialog!"),
+        ):
+            await set_shipping_options(test_bot, ad_cfg)
+
+
+class TestWantedShippingSelection:
+    """Tests for WANTED shipping path via set_shipping_form.
+
+    WANTED ads render shipping as a special-attribute combobox dropdown
+    (``<button role="combobox">``) rather than radio buttons. These tests
+    verify ``set_shipping_form`` directly with WANTED ad configurations.
+    """
+
+    @pytest.mark.asyncio
+    @pytest.mark.parametrize(
+        ("shipping_type", "expected_label"),
+        [("SHIPPING", "Versand möglich"), ("PICKUP", "Nur Abholung")],
+        ids = ["shipping", "pickup"],
+    )
+    async def test_wanted_shipping_selects_combobox_dropdown(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+        shipping_type:str,
+        expected_label:str,
+    ) -> None:
+        """WANTED ads should select shipping via button-combobox dropdown using VERSAND_COMBOBOX_SELECTOR."""
+        ad_cfg = Ad.model_validate(
+            base_ad_config
+            | {
+                "type": "WANTED",
+                "shipping_type": shipping_type,
+                "shipping_options": [],
+            }
+        )
+
+        combobox_btn = MagicMock()
+        combobox_btn.attrs = {"id": "babyausstattung.versand"}
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = combobox_btn) as mock_find,
+            patch.object(test_bot, "web_select_button_combobox", new_callable = AsyncMock) as mock_select_btn_combo,
+        ):
+            await set_shipping_form(test_bot, ad_cfg)
+
+        mock_find.assert_awaited_once_with(
+            By.CSS_SELECTOR,
+            VERSAND_COMBOBOX_SELECTOR,
+            timeout = test_bot.timeout("quick_dom"),
+        )
+        mock_select_btn_combo.assert_awaited_once_with(
+            "babyausstattung.versand",
+            expected_label,
+        )
+
+    @pytest.mark.asyncio
+    async def test_wanted_shipping_raises_when_combobox_not_found(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """WANTED shipping should fail with TimeoutError when the combobox button cannot be found."""
+        ad_cfg = Ad.model_validate(
+            base_ad_config
+            | {
+                "type": "WANTED",
+                "shipping_type": "SHIPPING",
+                "shipping_options": [],
+            }
+        )
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, side_effect = TimeoutError("combobox not found in DOM")),
+            patch.object(test_bot, "web_select_button_combobox", new_callable = AsyncMock),
+            pytest.raises(TimeoutError, match = "Failed to set shipping attribute for type 'SHIPPING'!"),
+        ):
+            await set_shipping_form(test_bot, ad_cfg)
+
+    @pytest.mark.asyncio
+    async def test_wanted_shipping_not_applicable_skips_combobox(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """WANTED ads with NOT_APPLICABLE shipping should skip the combobox entirely."""
+        ad_cfg = Ad.model_validate(
+            base_ad_config
+            | {
+                "type": "WANTED",
+                "shipping_type": "NOT_APPLICABLE",
+                "shipping_options": [],
+            }
+        )
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
+            patch.object(test_bot, "web_select_button_combobox", new_callable = AsyncMock) as mock_select_btn_combo,
+        ):
+            await set_shipping_form(test_bot, ad_cfg)
+
+        mock_find.assert_not_awaited()
+        mock_select_btn_combo.assert_not_awaited()
+
+    @pytest.mark.asyncio
+    async def test_wanted_shipping_raises_when_combobox_has_no_id(
+        self,
+        test_bot:KleinanzeigenBot,
+        base_ad_config:dict[str, Any],
+    ) -> None:
+        """WANTED shipping should fail with TimeoutError when the combobox button has no id attribute."""
+        ad_cfg = Ad.model_validate(
+            base_ad_config
+            | {
+                "type": "WANTED",
+                "shipping_type": "SHIPPING",
+                "shipping_options": [],
+            }
+        )
+
+        combobox_btn = MagicMock()
+        combobox_btn.attrs = {}  # No "id" key
+
+        with (
+            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = combobox_btn),
+            patch.object(test_bot, "web_select_button_combobox", new_callable = AsyncMock),
+            pytest.raises(TimeoutError, match = "Failed to set shipping attribute for type 'SHIPPING'!"),
+        ):
+            await set_shipping_form(test_bot, ad_cfg)
